wordpress: remove classRef
We used to support classRefs by name and namespace, but now that we have
strongly typed resource claims and default resource policies, the way to
have a portable workload is by using a default resource policy, which
means no classRef at all. This changeset updates the wordpress template
to use no classRefs for its claims.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>